+++ tests/testthat/test-read.R
@@ -1,8 +1,4 @@
-context("read")
-
-describe("read() R package datasets", {
-
-  it("reads package datasets when package= argument is provided", {
-    expect_is(read("urchin_bio", package = "data.io"), "data.frame")
-  })
+test_that("read() R package datasets", {
+  expect_s3_class(read("urchin_bio", package = "data.io"), "data.frame")
 })
+